arxivGPT/src/requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
dataclasses-json==0.6.7
exceptiongroup==1.2.2
faiss-cpu==1.10.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-text-splitters==0.3.8
langsmith==0.3.26
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.3.2
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
orjson==3.10.16
packaging==24.2
pillow==11.1.0
propcache==0.3.1
psutil==7.0.0
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyMuPDF==1.25.5
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.2.2
tqdm==4.67.1
transformers==4.51.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
yarl==1.19.0
zstandard==0.23.0
